+++ docs/changelog.rst
@@ -27,7 +27,7 @@ v1.1.1
   * add :func:`craftr.TargetBuilder.setdefault` method
   * add :data:`craftr.FrameworkJoin.defaults` member
   * add :func:`craftr.FrameworkJoin.iter_frameworks` method
-  * moved ``craftr.ext.compiler.BaseCompiler`` to :class:`craftr.ext.compiler._base.BaseCompiler`,
+  * moved ``craftr.ext.compiler.BaseCompiler`` to :class:`craftr.ext.compiler.base.BaseCompiler`,
     backwards compatible import exists
   * removed ``BaseCompiler.__getitem__()`` and ``~.__setitem__()``
   * add ``BaseCompiler.register_hook()``